requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
openai>=1.0.0
faiss-cpu>=1.8.0
numpy>=1.26.0
pandas>=2.2.0
rank-bm25>=0.2.2
tqdm>=4.66.0
httpx>=0.27.0
pyarrow>=16.0.0
